Delete existing clusters after unrecoverable errors (#63)
* Restructure auth and subscription packages

Reorganize ocm auth and subscription content to simplify package
organization

* Recognize unrecoverable discovery errors

Expose function in ocm package to check if the error returned might
be a temporary or is unrecoverable. An unrecoverable error means
something is wrong with the credentials supplied and should result
in deleting any previously discovered clusters

* Add functional tests for bad creds

Make sure clusters are deleted when credentials go invalid. Also
clean up if the backing secret is deleted.

* Add test to confirm ErrInvalidToken returned