Multi-namespace support enhancements (#68)
* Restrict naming to only accept 'discoveryconfig'

Only reconcile on discoveryconfigs with the name discoveryconfig,
thereby enforcing the one config per namespace rule.

* Update managed status across namespaces

This will ensure when a cluster changes managed status, discovered
clusters across all namespaces are updated.

* Update ManagedCluster CRD file

* Add multi-namespace functional tests

Introduce tests to validate functionality when the controller is
managing discoveryconfigs in multiple namespaces. Also ensures that
if a cluster becomes managed it updates discoveredclusters that may
be duplicated in more than one namespace.